小さなUIコンポーネントのmarginの除去、最上位コンポーネントにはpropsを持たせないよう修正
