+++ todoapp/src/components/atoms/Card/Card.styles.tsx
@@ -2,12 +2,11 @@
 import { Card } from '@mui/material';
 import { styled } from '@mui/system';
 
-export const StyledCard = styled(Card)({
+export const TodoItemCard = styled(Card)({
     display: 'flex',
     justifyContent: 'space-between',
     alignItems: 'center',
     padding: '10px 0',
     border: '2px solid #ddd',
     borderRadius: 10,
-    marginBottom: 20,
 });
