+++ todoapp/src/components/atoms/Container/Container.styles.tsx
@@ -2,7 +2,10 @@
 import { Container } from '@mui/material';
 import { styled } from '@mui/system';
 
-export const StyledContainer = styled(Container)({
+export const TodoListContainer = styled(Container)({
+    display: 'flex',
+    flexDirection: 'column',
+    gap: '10px',
     width: '80%',
     margin: '50px auto',
     padding: '20px',
