+++ todoapp/src/components/molecules/TodoItem/TodoItem.tsx
@@ -1,21 +1,19 @@
 import React from 'react';
 import type { Task } from '../../../types/task';
 import { StyledButton } from '../../atoms/Button/Button.styles';
-import { StyledCard } from '../../atoms/Card/Card.styles';
+import { TodoItemCard } from '../../atoms/Card/Card.styles';
 import { StyledTypography} from '../../atoms/Typography/Typography.styles';
 
-export const TodoItem: React.FC<Task & { index: number }> = (props) => {
-
-    const { id,title } = props;
+export const TodoItem: React.FC<Task & { index: number }> = ({ title }) => {
 
     return (
-        <StyledCard key={id}>
+        <TodoItemCard>
             <StyledTypography>
                 {title}
             </StyledTypography>
             <StyledButton>
                 削除
             </StyledButton>
-        </StyledCard>    
+        </TodoItemCard>    
     )
 }
